requirements.txt
pika >= 0.12.0
pykerberos >= 1.2.1